chore(storage): reorgnize code
